better alignment for ans

+++ pco/src/ans/mod.rs
@@ -11,7 +11,7 @@ mod spec;
 pub(crate) type AnsState = u32;
 // must be u16 or larger
 // should not be exposed in public API
-pub(crate) type Token = u16;
+pub(crate) type Token = u32;
 
 #[cfg(test)]
 mod tests {
